added enc msg

+++ EasyChat/Crypto_Manager.h
@@ -12,8 +12,8 @@ namespace Crypto_Manager{
 
 	std::string get_sha3_512_hash(std::string message);
 	std::string bytes_to_hex_string(const std::vector<uint8_t>& bytes);
-	std::vector<size_t> rsa_encrypt(std::string message);
-	std::string rsa_decrypt(std::vector<size_t> encrypted_message);
+	std::vector<double> rsa_encrypt(std::string message);
+	std::string rsa_decrypt(std::vector<double> encrypted_message);
 
 	double rsa_encrypt_char(char c);
 	char rsa_decrypt_char(double encrypted);
